player specific draw routine, colors, whatnot

+++ game/entities/entity.c
@@ -23,5 +23,5 @@ void update_entity(entity *e, float dt) {
 void draw_entity(entity *e, Texture2D tex) {
     // scale needs to be taken into account in setting up the origin for correct rotations
     Vector2 origin = {tex.width/2 * e->scale, tex.height/2 * e->scale};
-    DrawTexturePro(tex, (Rectangle){0, 0, tex.width, tex.height}, (Rectangle){e->x, e->y, tex.width*e->scale, tex.height*e->scale}, origin, e->rotation, WHITE);
+    DrawTexturePro(tex, (Rectangle){0, 0, (float)tex.width, (float)tex.height}, (Rectangle){e->x, e->y, tex.width*e->scale, tex.height*e->scale}, origin, e->rotation, WHITE);
 }
